pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ss-anim-mcp"
version = "0.1.0"
description = "MCP server for automated spritesheet to game animation conversion"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "I2g Team" }
]
keywords = ["mcp", "aseprite", "spritesheet", "animation", "game-dev"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "watchfiles>=0.21.0",
    "watchdog>=3.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
ss-anim-mcp = "ss_anim_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ss_anim_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
